+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # Change log for refreshVersions
 
+## Version 0.60.1 (2023-08-24)
+
+### Bug fix
+
+- Fix crash on projects that had no version catalog.
+
+### New dependency notations:
+
+<details>
+<summary><strong>Click to expand (25) </strong></summary>
+- `Google.android.play.appUpdateKtx`
+- `Google.android.play.appUpdate`
+- `Google.android.play.assetDeliveryKtx`
+- `Google.android.play.assetDelivery`
+- `Google.android.play.featureDeliveryKtx`
+- `Google.android.play.featureDelivery`
+- `Google.android.play.reviewKtx`
+- `Google.android.play.review`
+- `Google.horologist.annotations`
+- `Google.horologist.auth.composables`
+- `Google.horologist.auth.data.phone`
+- `Google.horologist.auth.data.watchOAuth`
+- `Google.horologist.auth.data`
+- `Google.horologist.auth.ui`
+- `Google.horologist.compose.material`
+- `Google.horologist.dataLayer.grpc`
+- `Google.horologist.dataLayer.phone`
+- `Google.horologist.dataLayer.watch`
+- `Google.horologist.healthComposables`
+- `Google.horologist.media3.audioOffload`
+- `Google.horologist.media3.logging`
+- `Google.horologist.media3.outputSwitcher`
+- `Google.horologist.networkAwareness.db`
+- `Google.horologist.networkAwareness.okHttp`
+- `Google.horologist.networkAwareness.ui`
+</details>
+
 ## Version 0.60.0 (2023-08-17)
 
 ### Full configuration cache support for the `refreshVersions` task